requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.11.7
pydantic-settings==2.1.0
python-multipart==0.0.6
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiohttp==3.9.1
requests==2.31.0
Pillow==11.3.0
